backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
pydantic
fastapi-cors
openai
google-generativeai